change namespace and class name from Amqp to RabbitMQ
