+++ src/Bedrock.Framework.Experimental/Protocols/RabbitMQ/IRabbitMQMessage.cs
@@ -3,7 +3,7 @@ using System.Buffers;
 using System.Collections.Generic;
 using System.Text;
 
-namespace Bedrock.Framework.Experimental.Protocols.Amqp
+namespace Bedrock.Framework.Experimental.Protocols.RabbitMQ
 {
     public interface IAmqpMessage
     {
