+++ src/Bedrock.Framework.Experimental/Protocols/RabbitMQ/Methods/ConnectionOpenOk.cs
@@ -4,30 +4,20 @@ using System.Buffers.Binary;
 using System.Collections.Generic;
 using System.Text;
 
-namespace Bedrock.Framework.Experimental.Protocols.Amqp.Methods
+namespace Bedrock.Framework.Experimental.Protocols.RabbitMQ.Methods
 {
-    public class ConnectionTune : MethodBase, IAmqpMessage
+    public class ConnectionOpenOk : MethodBase, IAmqpMessage
     { 
         public override byte ClassId => 10;
-        public override byte MethodId => 30;
-
-        public short MaxChannel { get; private set; }
-        public int MaxFrame { get; private set; }
-        public short HeartBeat { get; private set; }
+        public override byte MethodId => 41;
+        public ReadOnlyMemory<byte> Reserved1 { get; private set; }
 
         public bool TryParse(ReadOnlySequence<byte> input,  out SequencePosition end)
         {
             SequenceReader<byte> reader = new SequenceReader<byte>(input);            
             try
             {
-                reader.TryReadBigEndian(out short maxChannel);
-                reader.TryReadBigEndian(out int maxFrame);
-                reader.TryReadBigEndian(out short heartBeat);
-
-                MaxChannel = maxChannel;
-                MaxFrame = maxFrame;
-                HeartBeat = heartBeat;
-
+                Reserved1 = ProtocolHelper.ReadShortString(ref reader);
                 end = reader.Position;
                 return true;
             }
