+++ src/Bedrock.Framework.Experimental/Protocols/RabbitMQ/Methods/ConnectionTuneOk.cs
@@ -4,7 +4,7 @@ using System.Buffers.Binary;
 using System.Collections.Generic;
 using System.Text;
 
-namespace Bedrock.Framework.Experimental.Protocols.Amqp.Methods
+namespace Bedrock.Framework.Experimental.Protocols.RabbitMQ.Methods
 {
     public class ConnectionTuneOk : MethodBase, IAmqpMessage
     {
@@ -30,7 +30,7 @@ namespace Bedrock.Framework.Experimental.Protocols.Amqp.Methods
         public void Write(IBufferWriter<byte> output)
         {  
             var payloadLength = sizeof(ushort) + sizeof(uint) + sizeof(ushort) + MethodHeaderLength;
-            var buffer = output.GetSpan(AmqpMessageFormatter.HeaderLength + payloadLength + 1);
+            var buffer = output.GetSpan(RabbitMQMessageFormatter.HeaderLength + payloadLength + 1);
 
             WriteHeader(ref buffer, 0, payloadLength);
 
@@ -41,7 +41,7 @@ namespace Bedrock.Framework.Experimental.Protocols.Amqp.Methods
             BinaryPrimitives.WriteUInt16BigEndian(buffer.Slice(10), (ushort)this.HeartBeat);
             buffer[payloadLength]= (byte)FrameType.End;
 
-            output.Advance(AmqpMessageFormatter.HeaderLength + payloadLength + sizeof(byte));
+            output.Advance(RabbitMQMessageFormatter.HeaderLength + payloadLength + sizeof(byte));
         }
     }
 }
