+++ src/Bedrock.Framework.Experimental/Protocols/RabbitMQ/Methods/QueueDeclare.cs
@@ -4,7 +4,7 @@ using System.Buffers.Binary;
 using System.Collections.Generic;
 using System.Text;
 
-namespace Bedrock.Framework.Experimental.Protocols.Amqp.Methods
+namespace Bedrock.Framework.Experimental.Protocols.RabbitMQ.Methods
 {
     public class QueueDeclare : MethodBase, IAmqpMessage
     {
@@ -41,7 +41,7 @@ namespace Bedrock.Framework.Experimental.Protocols.Amqp.Methods
         public void Write(IBufferWriter<byte> output)
         {          
             var payloadLength = 6 + 1 + QueueName.Length + sizeof(byte) + MethodHeaderLength;
-            var buffer = output.GetSpan(AmqpMessageFormatter.HeaderLength + payloadLength + 1);
+            var buffer = output.GetSpan(RabbitMQMessageFormatter.HeaderLength + payloadLength + 1);
            
             WriteHeader(ref buffer, this.Channel, payloadLength);
 
@@ -61,7 +61,7 @@ namespace Bedrock.Framework.Experimental.Protocols.Amqp.Methods
             BinaryPrimitives.WriteUInt32BigEndian(buffer.Slice(7 + QueueName.Length + 5), 0);
             buffer[payloadLength] = (byte)FrameType.End;
 
-            output.Advance(AmqpMessageFormatter.HeaderLength + payloadLength + sizeof(byte));
+            output.Advance(RabbitMQMessageFormatter.HeaderLength + payloadLength + sizeof(byte));
         }        
     }
 }
