+++ src/Bedrock.Framework.Experimental/Protocols/RabbitMQ/Methods/QueueDeclareOk.cs
@@ -4,7 +4,7 @@ using System.Buffers.Binary;
 using System.Collections.Generic;
 using System.Text;
 
-namespace Bedrock.Framework.Experimental.Protocols.Amqp.Methods
+namespace Bedrock.Framework.Experimental.Protocols.RabbitMQ.Methods
 {
     public class QueueDeclareOk : MethodBase, IAmqpMessage
     {
